+++ src/main/java/com/razdeep/konsignapi/model/CollectionVoucherItem.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class CollectionVoucherItem {
 
-    Float amountCollected;
+    Double amountCollected;
 
     String bank;
